+++ resources/assets/js/app/components/MainNavigation/ForGuestUser.js
@@ -2,8 +2,7 @@ import React from 'react';
 import { Link, IndexLink } from 'react-router';
 import { localize } from 'react-localize-redux';
 
-import ChangeLocalization from '../ChangeLocalization';
-import ChangeCurrency from '../ChangeCurrency';
+import { ChangeCurrency, ChangeLocalization } from '../MenuDropdowns';
 
 class ForGuestUser extends React.Component {
