+++ resources/assets/js/app/components/MenuDropdowns/index.js
@@ -1,0 +1,9 @@
+import UserDropdown from './UserDropdown';
+import ChangeCurrency from './ChangeCurrency';
+import ChangeLocalization from './ChangeLocalization';
+
+export {
+    UserDropdown,
+    ChangeCurrency,
+    ChangeLocalization,
+};
